appimage/baseapp/requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.21
botocore==1.31.21
Django==4.2.4
django-appconf==1.0.5
django-browser-reload==1.11.0
django-compressor==4.4
django-filter==23.2
django-tailwind==3.6.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
psycopg2-binary==2.9.7
pytailwindcss==0.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
rcssmin==1.1.1
referencing==0.30.2
rjsmin==1.2.1
rpds-py==0.9.2
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
